% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine2CDFs.R
\name{combine2CDFs}
\alias{combine2CDFs}
\title{Combine 2 Cumulative Distances}
\usage{
combine2CDFs(cdf1, cdf2)
}
\arguments{
\item{cdf1}{a list of cumulative distance values.}

\item{cdf2}{a list of cumulative distance values.}
}
\description{
Computes the combined values for 2 cumulative distances.
}
\details{
The combined values are computed by linear interpolation.
}
\author{
Rui Camacho, Paula Raissa
}
