% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCCDFs.R
\name{combineCCDFs}
\alias{combineCCDFs}
\title{Conditional Combine Cumulative Distribuition Function}
\usage{
combineCCDFs(CCDFlist)
}
\arguments{
\item{CCDFlist}{a list of conditional cumulative distance values}
}
\description{
Combine an entire list of Conditioned Cum Dist Funcs.
}
\details{
The combined values are computed recursively.
}
\author{
Rui Camacho, Paula Raissa
}
