% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.anova.R
\name{ds.anova}
\alias{ds.anova}
\title{Fit an Analysis of Variance Model}
\usage{
ds.anova(formula = NULL, model, weights = NULL, learningrate = 0.01,
  dif = 1e-09, checks = FALSE, datasources = NULL)
}
\arguments{
\item{formula}{a character that can be coerced to an object of class \code{\link[stats]{formula}}. It is a symbolic
description of the model to be fitted.}

\item{model}{a character, the name of regression model.
If \code{model} is set to 'linear', computes Anova for a linear model;
if \code{model} is set to 'logistic', computes Anova for a logistic model;
if \code{model} is set to 'poissin', computes Anova for a Poisson model.}

\item{weights}{a character, the name of an optional vector of weights to be used in the fitting process.
Should be null or a numeric vector.}

\item{learningrate}{a numeric, controls how much we are adjusting the regression model.
It is an optional parameter. Should be set if the Anova will be computed for logistic ou Poisson model.}

\item{dif}{a numeric, controls the learning convergence.
It is an optional parameter. Should be set if the Anova will be computed for logistic ou Poisson model.}

\item{checks}{a boolean, if TRUE (default) checks that verify elements on the server side
such checks lengthen the run-time so the default is FALSE and one can switch these checks
on (set to TRUE) when faced with some error(s).}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
Returns the Anova table with the following components:
\item{Df}{degrees of freedon}
\item{Sum.Sq}{sum of squares}
\item{Mean.Sq}{mean of squares}
\item{F.value}{f-test}
\item{Pr(>F)}{p-value from f-statistic}
}
\description{
Computes the combined equations for analysis of variance.
}
\details{
The variation between and within groups for a one-way analysis of variance generalizes to \emph{model variation}
and \emph{residual variation} which partition the total variation \eqn{SSD[model] = \sum[i](y[i]-y*)^2}.
This can be applied only when the model contains an intercept.
The model is considered to be statistically significant if it can account for a large amount of variability in the response.
}
\section{Dependencies}{

\code{\link{getAnova}}
}

\author{
Paula Raissa Costa e Silva
}
