% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.condCDF.R
\name{ds.condCDF}
\alias{ds.condCDF}
\title{Conditional Cumulative Distributed Function}
\usage{
ds.condCDF(varName, catName, datasources)
}
\arguments{
\item{varName}{a character, the name of study variable. The data type must be a numeric.}

\item{catName}{a character, the name of categorical variable.}

\item{datasources}{a list of parameters to access files sytems or databases.}
}
\value{
The combined cumulative distribution function of all datasources.
}
\description{
Computes the distributed cumulative distribution function.
}
\details{
Reads each datasource (node) and compute the local cumulative distributed function for each data node.
If the datasources are files, these files must be in .csv formate.
}
\examples{
{
ds.cdf(varName, fileNames)
}

}
\author{
Rui Camacho, Paula Raissa
}
