% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.condPercentile.R
\name{ds.condPercentile}
\alias{ds.condPercentile}
\title{Distributed Conditional Percentile}
\usage{
ds.condPercentile(varName, catName, category, xValue, datasources)
}
\arguments{
\item{varName}{a character, the name of numerical variable.}

\item{catName}{a character, the name of categorical variable.}

\item{category}{the given category to calculate the corresponding percentile.}

\item{xValue}{the given value to calculate the corresponding percentile.}

\item{datasources}{a list of parameters to access files sytems or databases.}
}
\description{
Computes the distributed conditional percentile for a given study and categorical variables.
}
\details{
Read each datasource (node) and compute the combined percentile according to categories.
}
\section{Dependencies}{

\code{\link{getSamplesCounting}}
\code{\link{getPercentile}}
}

\examples{
{
fileNames <- c("juul.csv", "juul2.csv")
varName <- "age"
catName <- "sex"
ds.CondPercentile(varName, catName, 'f', 10, fileNames)
}
}
\author{
Rui Camacho, Paula Raissa
}
