% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.fTest.R
\name{ds.fTest}
\alias{ds.fTest}
\title{Distributed F Test to Compare Two Variances}
\usage{
ds.fTest(x = NULL, y = NULL, ratio = 1, conf.level = 0.95,
  alternative = c("two.sided", "less", "greater"), datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a numeric vector of data values.}

\item{y}{a character, the name of a numeric vector of data values.}

\item{ratio}{the hypothesized ratio of the population variances of x and y.}

\item{conf.level}{confidence level for the returned confidence interval.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default),
"greater" or "less". You can specify just the initial letter.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
A list containing the following components:
\item{statistic}{the value of the F test statistic.}
\item{parameter}{the degrees of the freedom of the F distribution of the test statistic.}
\item{p.value}{the p-value of the test.}
\item{conf.int}{a confidence interval for the ratio of the population variances.}
\item{estimate}{the ratio of the sample variances of x and y.}
\item{null.value}{the ratio of population variances under the null.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{the character string "F test to compare two variances".}
}
\description{
Performs an F test to compare the variances of two samples from normal populations.
}
\details{
The null hypothesis is that the ratio of the variances of the populations from which x and y were drawn, or in
the data to which the linear models x and y were fitted, is equal to ratio.
}
\section{Dependencies}{

\code{\link{getLength}}, \code{\link{getVariance}}
}

\examples{
{
ds.fTest(x, y)
}

}
\author{
Paula Raissa Costa e Silva
}
