% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.geometricMean.R
\name{ds.geometricMean}
\alias{ds.geometricMean}
\title{Geometric Mean}
\usage{
ds.geometricMean(x = NULL, naRm = TRUE, zero.propagate = FALSE,
  datasources = NULL)
}
\arguments{
\item{x}{a character, the name of study variable.}

\item{naRm}{a boolean, if TRUE (default) remove the missing values during the computation on the server side;
if FALSE does not remove the missing values.}

\item{zero.propagate}{a boolean, if TRUE (default) remove all zeros during the computation on the server side;
if FALSE does not remove the zero values.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the geometric mean for a given study variable.
}
\details{
Formally the geometric mean is the \emph{i\link{n}} root of the \emph{n} numbers products, or \emph{exp} to the mean log of \emph{x}.
Each data node computes the sum of logs, and the central node computes the arithmetic mean and its exponentiation.
}
\section{Dependencies}{

\code{\link{getGm}}, \code{\link{getLength}}
}

\examples{
{
gm <- ds.geometricMean('D$birth_weight')
}
}
\author{
Paula Raissa Costa e Silva
}
