% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.histogram.R
\name{ds.histogram}
\alias{ds.histogram}
\title{Combined Histogram}
\usage{
ds.histogram(x = NULL, intervalList = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of study variable.}

\item{datasources}{a list of parameters to access files sytems or databases.}

\item{listValues}{a list of values to compute the frequency table.}
}
\value{
return a numeric value.
}
\description{
Computes the combined frequecy table for each value given in a list of intervals.
}
\details{
Return the frequency value for each given interval.
For a intervals list \link{V1, V2, V3, ...}, it computes the frequency list \link{F1, F2, F3, ...}, where F1 is the number of values <= V1;
F2 is the number of values where V1 > x <= V2, etc.
}
\section{Dependencies}{

\code{\link{getHistogram}}
}

\examples{
{
fileNames <- list(file=c("file1.csv", "file2.csv", "file3.csv"))
ds.histogram(x='idademae', intervalList=c(20,40,50,30,20,30), datasources=fileNames)
}
}
\author{
Paula Raissa Costa e Silva
}
