% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.kurtosis.R
\name{ds.kurtosis}
\alias{ds.kurtosis}
\title{Kurtosis}
\usage{
ds.kurtosis(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a numerical vector}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the kurtosis of a given vector.
}
\details{
Considering \eqn{x[i]} for elements of \eqn{x} and \eqn{x*} for their mean, so \eqn{m[r] = \sum[i](x[i]-x*)^r/n}
is the sample moments of order r.
Which \eqn{\sum[i](x[i]-x*)^r} is computed at each data node and combined by this function in the central node, also called client-side.

It is a wrapper for the server side function \code{\link{getKurtosis}}.
}
\section{Dependencies}{

\code{\link{getKurtosis}}, \code{\link{ds.arMean}}, \code{\link{ds.length}}
}

\examples{
{
kurt <- ds.kurtosis('D$birth_weight')
}
}
\author{
Paula Raissa Costa e Silva
}
