% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.length.R
\name{ds.length}
\alias{ds.length}
\title{Length of an Object}
\usage{
ds.length(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the sample length of a study variable.
}
\details{
If a vector is shortened, extra values are discarded and when a vector is lengthened,
it is padded out to its new length with NAs (nul for raw vectors).

It is a wrapper for the server side function \code{\link{getLength}}.
}
\section{Dependencies}{

\code{\link{getLength}}
}

\examples{
{
size <- ds.length('D$birth_weight')
}

}
\author{
Paula Raissa Costa e Silva
}
