% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.max.R
\name{ds.max}
\alias{ds.max}
\title{Maximum}
\usage{
ds.max(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
If type='split', returns a numeric list.
If type='combine', return a numeric value.
}
\description{
Computes the maximum value among data nodes or for each data node.
}
\details{
Return the maximum of all the values present in their arguments, as integer if all are logical or integer,
as double if all are numeric, and character otherwise.

The maximum of a numeric empty set is +Inf which ensures transitivity.
For numeric x ds.max(x) == -Inf whenever ds.length(x) == 0 (after removing missing values if requested).

By definition the max of a numeric vector containing an NaN is NaN, except that the max of any vector containing
an NA is NA even if it also contains an NaN. Note that ds.max(NA, Inf) == NA even though the maximum would be Inf whatever
the missing value actually is.

Character versions are sorted lexicographically. The max of an empty character vector is defined to be character NA.

It is a wrapper for the server side function \code{\link{getMax}}.
}
\section{Dependencies}{

\code{\link{getMax}}
}

\examples{
{
max <- ds.max('D$birth_weight')
}
}
\author{
Paula Raissa Costa e Silva
}
