% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.min.R
\name{ds.min}
\alias{ds.min}
\title{Minimum}
\usage{
ds.min(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the minimum value among data nodes or for each data node.
}
\details{
Return the minimum of all the values present in their arguments, as integer if all are logical or integer,
as double if all are numeric, and character otherwise.

The minimum of a numeric empty set is +Inf which ensures transitivity.
For numeric x ds.min(x) == +Inf whenever ds.length(x) == 0 (after removing missing values if requested).

By definition the min of a numeric vector containing an NaN is NaN, except that the min of any vector containing
an NA is NA even if it also contains an NaN. Note that ds.min(NA, Inf) == NA even though the minimum would be Inf whatever
the missing value actually is.

Character versions are sorted lexicographically. The min of an empty character vector is defined to be character NA.

It is a wrapper for the server side function \code{\link{getMin}}.
}
\section{Dependencies}{

\code{\link{getMin}}
}

\examples{
{
ds.min('D$birth_weight')
}
}
\author{
Paula Raissa Costa e Silva
}
