% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.modelCheckings.R
\name{ds.modelCheckings}
\alias{ds.modelCheckings}
\title{Summarizing Regression Models}
\usage{
ds.modelCheckings(regression = NULL, formula = NULL, weight = NULL,
  datasources = NULL)
}
\arguments{
\item{regression}{an object of regression model.}

\item{formula}{a string character to be transformed as an object of class \code{formula}.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
Returns a list with the following components:
\item{residuals}{the weighted residuals, the usual residuals rescaled by the square root of the weights specified in the call to lm.}
\item{residuals.quantis}{the residuals quartiles}
\item{coefficients}{the matrix of coefficients, standard errors, z-values and p-values. Aliased coefficients are omitted.}
\item{sigma}{the square root of the estimated variance of the random error.}
\item{df}{degrees of freedom, a 3-vector (p, n-p, p*), the first being the number of non-aliased coefficients, the last being the total number of coefficients.}
\item{r.squared}{\eqn{R^2}, the fraction of variance explained by the model
\deqn{R^2 = 1 - Sum(R[i]^2) / Sum((y[i]-y*)^2)},
where \eqn{y*} is the mean of y\link{i} if there is an intercept and zero otherwise.}
\item{adj.r.squared}{the above \eqn{R^2} statistic 'adjusted', penalizing for higher \emph{p}.}
\item{fstatistic}{(for models including non-intercept terms) a 3-vector with the value of the F-statistic with its numerator and denominator degrees of freedom.}
\item{pvalue}{p-value from f-statistic.}
\item{n}{the sample size.}
\item{cov.unscaled}{the unscaled (dispersion = 1) estimated covariance matrix of the estimated coefficients.}
}
\description{
Combined computations for model checking.
}
\details{
Combines the computation of the coefficients, standard errors, etc.
The coefficients component of the result gives the estimated coefficients and their estimated standard errors,
together with their ratio. This third column is labelled t ratio if the dispersion is estimated, and z ratio if the
dispersion is known (or fixed by the family). A fourth column gives the two-tailed p-value corresponding to the t or
z ratio based on a Student t or Normal reference distribution. (It is possible that the dispersion is not known and
there are no residual degrees of freedom from which to estimate it. In that case the estimate is NaN.)
}
\section{Dependencies}{

\code{\link{modelCheckingDS}}
}

\examples{
{
linear <- ds.linear('D$maternal_age~D$birth_weight')
logistic <- ds.logistic('menarche ~ D$age')
poisson <- ds.poisson('D$y ~ D$x')
ds.summary(linear)
ds.summary(logistic)
ds.summary(poisson)
}
}
\author{
Paula Raissa Costa e Silva
}
