% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.na.R
\name{ds.na}
\alias{ds.na}
\title{Missing Values of Study Variable}
\usage{
ds.na(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the missing values of a given study variable.
Count all values set as NA or another missing value indicator.
}
\details{
It is a wrapper for the server side function \code{\link{getNumNa}}.
}
\section{Dependencies}{

\code{\link{getNumNa}}
}

\examples{
{
na <- ds.na('D$birth_weight')
}

}
\author{
Paula Raissa Costa e Silva
}
