% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.percentile.R
\name{ds.percentile}
\alias{ds.percentile}
\title{Distributed Percentile}
\usage{
ds.percentile(varName = NULL, formula = NULL, xValue = NULL,
  cVar = NULL, relation = NULL, threshold = NULL, relation2 = NULL,
  threshold2 = NULL)
}
\arguments{
\item{xValue}{the given value to calculate the corresponding percentile.}

\item{x}{a character, the name of study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\description{
Computes the distributed percentile for a given study variable.
}
\details{
Read each data file (node) and compute the combined percentile value.
}
\section{Dependencies}{

\code{\link{getSamplesCounting}}
\code{\link{getPercentile}}
}

\examples{
{
ds.percentile(varName, 10, fileNames)
}
}
\author{
Rui Camacho, Paula Raissa
}
