% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.poisson.R
\name{ds.poisson}
\alias{ds.poisson}
\title{Poisson Regression}
\usage{
ds.poisson(formula, learningrate = 0.01, dif = 1e-09, checks = FALSE,
  datasources = NULL)
}
\arguments{
\item{formula}{a character that can be coerced to an object of class \code{\link[stats]{formula}}. It is a symbolic
description of the model to be fitted.
The Newton Raphson Method is apllied to compute the Poisson coefficients.}

\item{learningrate}{a numeric, controls how much we are adjusting the regression model.}

\item{dif}{a numeric, controls the learning convergence.}

\item{checks}{a boolean, if TRUE (default) checks that verify elements on the server side
such checks lengthen the run-time so the default is FALSE and one can switch these checks
on (set to TRUE) when faced with some error(s).}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
Returns a list with the following components:
\item{call}{the model formula.}
\item{coefficients}{a vector of poisson regression coefficients.}
\item{xtxw}{a data matrix, the \emph{Hessian} matrix.}
\item{xtyp}{a data matrix, that integrates the computation of derivatives.}
}
\description{
Computes Poisson regression model.
It can be used to fit univariate, multivariate and weighted Poisson models.
}
\details{
It is a wrapper for the client side function \code{\link{getDerivative}},
setting the model \code{family} as 'poisson'.
}
\section{Dependencies}{

\code{\link[distStatsC]{getDerivative}}
}

\examples{
{
ds.poisson('D$y ~ D$x')
}
}
\seealso{
Other regressions: \code{\link{ds.linear}},
  \code{\link{ds.logistic}}
}
\author{
Paula Raissa Costa e Silva
}
\concept{regressions}
