% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.prod.R
\name{ds.prod}
\alias{ds.prod}
\title{Product of study variable elements}
\usage{
ds.prod(x = NULL, naRm = FALSE, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of study variable.}

\item{naRm}{a boolean, if TRUE (default) remove the missing values during the computation on the server side;
if FALSE does not remove the missing values.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
It computes the product of all values in its arguments.
}
\details{
It is a wrapper for the server side function \code{\link{getProd}}.
}
\section{Dependencies}{

\code{\link{getProd}}
}

\examples{
prod <- ds.prod('D$maternal_age')

}
\author{
Paula Raissa Costa e Silva
}
