% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.sd.R
\name{ds.sd}
\alias{ds.sd}
\title{Standard Deviation}
\usage{
ds.sd(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the standard deviation for a study variable.
}
\details{
Uses denominator n-1.

The standard deviation of a zero-length vector (after removal of NAs if na.rm = TRUE) is not defined and gives an error.
The standard deviation of a length-one vector is NA.
}
\section{Dependencies}{

\code{\link[distStatsC]{ds.var}}
}

\examples{
{
sd <- ds.sd('D$birth_weight')
}
}
\author{
Paula Raissa Costa e Silva
}
