% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.sem.R
\name{ds.sem}
\alias{ds.sem}
\title{Standard Error of the Mean}
\usage{
ds.sem(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the variance for a study variable.
}
\details{
By default, computes the sample variance.
}
\section{Dependencies}{

\code{\link{getLength}}, \code{\link{getVariance}}
}

\examples{
{
var <- ds.var('D$birth_weight')
}

}
\author{
Paula Raissa Costa e Silva
}
