% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.skewness.R
\name{ds.skewness}
\alias{ds.skewness}
\title{Skewness}
\usage{
ds.skewness(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of the study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the skewness.
}
\details{
Considering x.i for elements of x and mu for their mean, so m.r = sum{i}(xi-mu)^r/n is the sample moments of order r.
Which sum{i}(xi-mu)^r is computed at each data node and combined by this function in the central node, also called client-side.

It is a wrapper for the server side function \code{\link{getSkewness}}.
}
\section{Dependencies}{

\code{\link{getSkewness}}
}

\examples{
skew <- ds.skewness('D$birth_weight')

}
\author{
Paula Raissa Costa e Silva
}
