% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.sum.R
\name{ds.sum}
\alias{ds.sum}
\title{Sum of Study Variable Elements}
\usage{
ds.sum(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a study variable.}

\item{datasources}{a list of a list of data connection parameters.
If it is a DataSHIELD connection, datasources is a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.
If the data is csv files, datasources is a list of paths and filenames.
If the data is a sql connection, datasources is a list of conection, database name, variables and query.}
}
\value{
return a numeric value.
}
\description{
Computes the sum of values from a given study variable.
}
\details{
It is a wrapper for the server side function \code{\link{getSum}}.
}
\section{Dependencies}{

\code{\link{getSum}}
}

\examples{
ds.sum('D$birth_weight')

}
\author{
Paula Raissa Costa e Silva
}
