% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.sumOfSquares.R
\name{ds.sumOfSquares}
\alias{ds.sumOfSquares}
\title{Sum of Squares}
\usage{
ds.sumOfSquares(x = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a study variable.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
return a numeric value.
}
\description{
Computes the sum of squares for a given study variable.
}
\details{
It is a wrapper for the server side function \code{\link{getSquaredSum}}.
}
\section{Dependencies}{

\code{\link{getSquaredSum}}
}

\examples{
ds.sumOfSquares(x='D$birth_weight')

}
\author{
Paula Raissa Costa e Silva
}
