% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tTest.R
\name{ds.tTest}
\alias{ds.tTest}
\title{Distributed Student's t-Test}
\usage{
ds.tTest(x, y = NULL, mu = 0, paired = FALSE, conf.level = 0.95,
  var.equal = FALSE, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a numeric vector of data values.}

\item{y}{x a character, the name of a optional numeric vector of data values.}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{conf.level}{confidence level of the interval.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal.
If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
A list containing the following components:
\item{statistic}{the value of the t-statistic.}
\item{parameter}{the degrees of freedom for the t-statistic.}
\item{p.value}{the p-value for the test.}
\item{conf.int}{a confidence interval for the mean appropriate to the specified alternative hypothesis.}
\item{estimate}{the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.}
\item{null.value}{the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample test or a two-sample test.}
\item{method}{a character string indicating what type of t-test was performed.}
}
\description{
Computes one and two sample t-tests on vectors of data
}
\details{
If paired is TRUE then both x and y must be specified and they must be the same length. Missing values are silently
removed (in pairs if paired is TRUE). If var.equal is TRUE then the pooled estimate of the variance is used. By default, if
var.equal is FALSE then the variance is estimated separately for both groups and the Welch modification to the degrees of freedom is used.
}
\section{Dependencies}{

\code{\link{getLength}}, \code{\link{getVariance}}
}

\examples{
{
ds.tTest(x)
ds.tTest(x, y)
}

}
\author{
Paula Raissa Costa e Silva
}
