% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.wilcoxTest.R
\name{ds.wilcoxTest}
\alias{ds.wilcoxTest}
\title{Distributed Wilcoxon Rank Sum and Signed Rank Tests}
\usage{
ds.wilcoxTest(x, y = NULL, alternative = c("two.sided", "less",
  "greater"), mu = 0, paired = FALSE, exact = NULL, correct = TRUE,
  conf.int = FALSE, conf.level = 0.95, datasources = NULL)
}
\arguments{
\item{x}{a character, the name of a numeric vector of data values.}

\item{y}{a character, the name of an optional numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".
You can specify just the initial letter.}

\item{mu}{anumber specifying an optional parameter used to form the null hypothesis.}

\item{paired}{a logical indicating whether you want a paired test.}

\item{exact}{a logical indicating whether an exact p-value should be computed.}

\item{correct}{a logical indicating whether to apply continuity correction in the normal approximation for the p-value.}

\item{conf.int}{a logical indicating whether a confidence interval should be computed.}

\item{conf.level}{confidence level of the interval.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
A list containg the following components:
\item{statistic}{the value of the test statistic with a name describing it.}
\item{parameter}{the parameter(s) for the exact distribution of the test statistic.}
\item{p.value}{the p-value for the test.}
\item{null.value}{the location parameter mu.}
\item{conf.int}{a confidence interval for the location parameter. If argument conf.int = TRUE.}
\item{estimate}{an estimate of the location parameter. If argument conf.int = TRUE.}
}
\description{
Computes one and two-sample Wilcoxon tests on vectors of data.
}
\details{
The formula interface is only applicable for the 2-sample tests.
f only x is given, or if both x and y are given and paired is TRUE, a Wilcoxon signed rank test of the null that the
distribution of x (in the one sample case) or of x - y (in the paired two sample case) is symmetric about mu is performed.
Otherwise, if both x and y are given and paired is FALSE, a Wilcoxon rank sum test
is carried out. In this case, the null hypothesis is that the distributions of x and y differ by a location shift of mu and
the alternative is that they differ by some other location shift (and the one-sided alternative "greater" is that x is shifted to the right of y).
}
\section{Dependencies}{

\code{\link{getLength}}, \code{\link{getVariance}}
}

\examples{
{
ds.wilcoxTest(x)
ds.wilcoxTest(x, y)
}

}
\author{
Paula Raissa Costa e Silva
}
