% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.xOfCondPercentile.R
\name{ds.xOfCondPercentile}
\alias{ds.xOfCondPercentile}
\title{X Value for Conditional Percentile}
\usage{
ds.xOfCondPercentile(varName, varCategory, category, percentile,
  datasources)
}
\arguments{
\item{varName}{a character, the name of study variable. It must be a numerical vector.}

\item{varCategory}{a character, the name of the categorical variable.}

\item{category}{a character, the category value.}

\item{percentile}{a numeric, the percentile value.}

\item{datasources}{a list of parameters to access files sytems or databases.}
}
\description{
Computes the X value for a given conditional percentile in distributed datasources
}
\details{
Read each data file (node) and compute the combined X value for a given percentile, according to a given category variable and name.
}
\section{Dependencies}{

\code{\link{getSamplesCounting}}
\code{\link{getPercentile}}
}

\examples{
{
ds.xOfCondPercentile('V3', 'gender', 'male', 0.1, datasources)
ds.xOfCondPercentile('V3', 'gender', 'female', 0.65, datasources)
}

}
\author{
Paula Raissa Costa e Silva
}
