% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.xOfPercentile.R
\name{ds.xOfPercentile}
\alias{ds.xOfPercentile}
\title{X Value for Percentile}
\usage{
ds.xOfPercentile(varName, percentile, datasources)
}
\arguments{
\item{varName}{a character, the name of study variable.}

\item{percentile}{a numeric, the given percentile to calculate the corresponding value.}

\item{datasources}{a list, parameters to access files sytems or databases.}
}
\description{
Computes the X value for a given percentile in distributed datasources
}
\details{
Read each data file (node) and compute the combined X value for a given percentile.
}
\section{Dependencies}{

\code{\link{getSamplesCounting}}
\code{\link{getPercentile}}
}

\examples{
{
ds.xOfPercentile('V2', 0.10, datasources)
}

}
\author{
Paula Raissa Costa e Silva
}
